removed parkApi.js; added commetns to searchRoutes.js; added getPark and router.get method to render park information in parkRoutes
